New ecss functionality for upsat_control (#136)
* Fix argument formatting in receiver.py

Seems like a regression in c466036, the way we format the arguments
to the demodulation_cmd and decoding_cmd changed. With this, my
default configuration bailed on attempting arguments not as strings
(the gain maybe?). In this commit I restore the argument formatting
we had previously, which does work with the same config.

* Prevent multiple jobs running concurrently

Previously, for some unknown reason when the observation task in
the client started, the client would attempt 2 instances of rtl_fm
and sometimes 2 instances of oggenc. In the case of a second
instance of rtl_fm it dies out quickly because the USB device is
already taken. In the case of a second oggenc, it will result in a
file with 0 bites overwriting the file with the actual data, which
the client will stash in /data/incomplete/.

I do not know why appscheduler starts 2 instances, I've checked
the jobs db and there is truly only one job per task. Seems to be
a common issue with appscheduler for differing reasons though. A
quick fix for this is to limit the number of job instances to 1.
There should be no reason why we would run multiple decoding_cmd's
or multiple demodulation_cmd's at the same time. In testing, I
have not reproduced the problem after dropping this setting to 1.

* Adding scheduling service: ui elements

* Added support for signed numbers conversion, updated adcs HK

* Added scheduling sid in MS, updated file uplink structure

* Updated data array handling

* Updated MS report, in order to include the head, tail fname

* Fixed display for sch

* Fixed event unitialized error

* Fixed utc data var not created, changed datepicker id to time & sch

* Fixed HK report variable, added app id to str

* Added ext wod time view

* Added qb50 to utc conversion

* Added subservices in time

* Added int to str conv

* Changed to dict and made checks if the dict exists

* Added task times

* Fix flake8 issues.