fix killing of gnuradio subprocesses
With Popen using shell=true, 2 processes are spawned off for the gnuradio listener. Through the terminate() method the parent process is killed but that does not always mean the child process is also killed, leaving a gnuradio script holding on to the SDR.

This change kills the entire process group instead of just the parent shell process. SIGKILL works fine in this case, no need for SIGTERM.

fixes satnogs/gr-satnogs#63
resolves satnogs/satnogs-client#253