Resolve threshold issues with rotators that report azimuth overwind positions as values outside 0-360 degrees
Some rotators (for example Spid Rot2Prog rotators) report positions in an overwind region
(i.e. out of the usual 0-360 degree range) as either negative values, or positive values
outside the 0-360 range. With the current approach to rotator movement thresholding, a value
outside the 0-360 range will always cause the threshold comparison to fail, which means the
rotator is almost continuously being commanded. This can result in premature failure of motor
relays.
This change causes the reported azimuth value to be wrapped to 0-360, which allows the threshold
function to perform correctly.

Signed-off-by: Mark Jessop <vk5qi@rfhead.net>